+++ frontend/src/dialogs/ProjectDialog.vue
@@ -133,9 +133,8 @@ limitations under the License.
 import { mapActions, mapGetters, mapState } from 'vuex'
 import { maxLength, required } from 'vuelidate/lib/validators'
 import { resourceName, unique, noStartEndHyphen, noConsecutiveHyphen } from '@/utils/validators'
-import { getValidationErrors, setInputFocus, isServiceAccount, compileMarkdown } from '@/utils'
+import { getValidationErrors, setInputFocus, isServiceAccount, compileMarkdown, getProjectDetails } from '@/utils'
 import { errorDetailsFromError, isConflict, isGatewayTimeout } from '@/utils/error'
-import { getProjectDetails } from '@/utils/projects'
 import cloneDeep from 'lodash/cloneDeep'
 import get from 'lodash/get'
 import map from 'lodash/map'
